Blue Team/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "blue-team-ai"
version = "0.1.0"
description = "Enterprise Copilot Blue Team agentic AI application with layered defenses and observability."
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [
  {name = "Blue Team"}
]
dependencies = [
  "typer>=0.12.3",
  "rich>=13.7.1",
  "pydantic>=2.7.4",
  "pydantic-settings>=2.4.0",
  "langchain-core>=0.2.35",
  "langchain-community>=0.2.6",
  "langchain-openai>=0.1.7",
  "openai>=1.37.0",
  "chromadb>=0.5.5",
  "sentence-transformers>=3.0.1",
  "scikit-learn>=1.4.2",
  "numpy>=1.26.4",
  "pandas>=2.2.2",
  "tiktoken>=0.7.0",
  "python-dotenv>=1.0.1",
  "structlog>=24.1.0",
  "fastapi>=0.111.0",
  "uvicorn[standard]>=0.29.0",
  "httpx>=0.27.0",
  "cachetools>=5.3.3",
  "sqlmodel>=0.0.16",
  "alembic>=1.13.2",
  "passlib[bcrypt]>=1.7.4",
  "cryptography>=42.0.8",
  "pyjwt[crypto]>=2.8.0",
  "opentelemetry-sdk>=1.25.0",
  "opentelemetry-exporter-otlp>=1.25.0"
]

[project.optional-dependencies]
dev = [
  "pytest>=8.2.2",
  "pytest-asyncio>=0.23.6",
  "pytest-cov>=5.0.0",
  "ruff>=0.5.0",
  "mypy>=1.10.0",
  "types-requests>=2.32.0.20240622"
]

[project.urls]
homepage = "https://example.com/blue-team-ai"
repository = "https://github.com/example/blue-team-ai"
issues = "https://github.com/example/blue-team-ai/issues"

